refactor(playground): split execution tree visualizer
