+++ packages/agent-playground/src/components/playground/execution-tree-visualizer/constants.ts
@@ -1,0 +1,1 @@
+export const MS_PER_SECOND = 1000;
