+++ packages/agent-playground/src/components/playground/execution-tree-visualizer/index.tsx
@@ -1,0 +1,2 @@
+export { ExecutionTreeVisualizer, default } from './execution-tree-visualizer';
+export type { IExecutionTreeVisualizerProps } from './types';
