+++ packages/agent-playground/src/components/playground/execution-tree-visualizer/types.ts
@@ -1,0 +1,18 @@
+import type { IPlaygroundBlockCollector } from '../../../lib/playground/block-tracking/block-collector';
+import type { IRealTimeBlockMessage } from '../../../lib/playground/block-tracking/types';
+
+export interface IExecutionTreeNode {
+  block: IRealTimeBlockMessage;
+  children: IExecutionTreeNode[];
+  level: number;
+}
+
+export interface IExecutionTreeVisualizerProps {
+  blockCollector: IPlaygroundBlockCollector;
+  showDebug?: boolean;
+  showProgress?: boolean;
+  autoExpand?: boolean;
+  onBlockSelect?: (block: IRealTimeBlockMessage) => void;
+  selectedBlockId?: string;
+  blockFilter?: (block: IRealTimeBlockMessage) => boolean;
+}
